+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Magento2 / Create password endpoint - @fifciu (#366)
 - API and Platform comments for all the REST endpoints added with the links to the official docs - @pkarw
 - Added global logger which you can change the instance of to use every logger you want to use - @resubaka (#24)
 - Added better typescript annotation/new types - @resubaka (#24)